refactor(arl): add and edit buttons
