+++ src/components/delete-button-with-dialog.tsx
@@ -35,12 +35,11 @@ export function DeleteButtonWithDialog({
   resource: Resource;
   id: string | number;
   itemName?: string;
-  variant?: VariantProps<typeof buttonVariants>["variant"];
-}) {
+} & VariantProps<typeof buttonVariants>) {
   const [isDialogOpen, setIsDialogOpen] = useState(false);
   const router = useRouter();
 
-  const sanitizedId = sanitizeId(String(id));
+  const sanitizedId = sanitizeId(id);
   const { mutateAsync, isPending, isSuccess } = useMutationWrapper<
     MessageResponse,
     null
